requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Detection
ultralytics>=8.0.0

# VLM
llava @ git+https://github.com/haotian-liu/LLaVA.git

# Data processing
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

# Experiment management
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0

# Auto-tuning
optuna>=3.4.0
optuna-dashboard>=0.13.0

# Evaluation
pycocoevalcap>=1.2
nltk>=3.8.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0
typer>=0.9.0
pyyaml>=6.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
